abstract the query command codec
